exo: ask confirmation before delete CMDs (#216)

+++ cmd/exo/cmd/sshkey_delete.go
@@ -1,12 +1,14 @@
 package cmd
 
 import (
+	"fmt"
+
 	"github.com/exoscale/egoscale"
 	"github.com/spf13/cobra"
 )
 
 // deleteCmd represents the delete command
-var deleteCmd = &cobra.Command{
+var sshkeyDeleteCmd = &cobra.Command{
 	Use:     "delete <name>",
 	Short:   "Delete ssh keyPair",
 	Aliases: gDeleteAlias,
@@ -14,6 +16,18 @@ var deleteCmd = &cobra.Command{
 		if len(args) < 1 {
 			return cmd.Usage()
 		}
+
+		force, err := cmd.Flags().GetBool("force")
+		if err != nil {
+			return err
+		}
+
+		if !force {
+			if !askQuestion(fmt.Sprintf("sure you want to delete %q ssh keyPair", args[0])) {
+				return nil
+			}
+		}
+
 		res, err := deleteSSHKey(args[0])
 		if err != nil {
 			return err
@@ -33,5 +47,6 @@ func deleteSSHKey(name string) (string, error) {
 }
 
 func init() {
-	sshkeyCmd.AddCommand(deleteCmd)
+	sshkeyDeleteCmd.Flags().BoolP("force", "f", false, "Attempt to remove ssh keyPair without prompting for confirmation")
+	sshkeyCmd.AddCommand(sshkeyDeleteCmd)
 }
